refactor(ai): extract heavyMaintenanceLeasePrimitives (Neo/Base-free) for subprocess lease consumers (#14204)
Extract the pure lease functions from HeavyMaintenanceLeaseService.mjs into a Neo/Base-free heavyMaintenanceLeasePrimitives.mjs (via git mv — blame preserved). The thin service imports the primitives, re-exports them (export *), and keeps the Neo class. All 9 existing importers are unaffected (8 named-function imports + ai/services.mjs's default class); the new module loads WITHOUT Neo/Base, so the kbSync VectorService subprocess can import shouldYieldHeavyMaintenanceLease and the lock functions without pulling in the orchestrator class stack (the #14186 layering fix).

Verified: both files pass node --check; the primitives import Neo/Base-free; all 19 HeavyMaintenanceLeaseService specs pass (behavior preserved through the re-export). Block-alignment --fix corrected 2 pre-existing misaligned lines surfaced by the rename.
